requirements.txt
click==8.0.1
colorama==0.4.4
cycler==0.10.0
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.20.3
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.6.3
six==1.16.0
Werkzeug==2.0.1
